8.Flask BluePrints/requirements.txt
Flask==2.3.3
pymongo==4.5.0
PyJWT==2.8.0
python-dotenv==1.0.0
Flask-Limiter==3.5.0
Flask-Cors==4.0.0
Werkzeug==2.3.7
marshmallow==3.20.1